Move AvatarRow outside of dashboard
