+++ frontend/src/components/generic/AvatarRow.vue
